I have restructured the program considerably scinse the java FX library demands that the class inheriting from "Application" must be initialized from within itselfe. Thus the new main class is the "GameFrame" class is the new main class.
The "LoadMap" class is a singgelton that is supposed to load the map (all of the objects) from a file. Right now it just instantiatiates a few
hardcoded objects to test that everything is working.

The "StaticSquare" class is the new name of the former "Ground" class. A lot of work has been done in said class but it will probably be split in
furure commits to a parent and a child to be able to create a dynamic square as well.

Right now the program spawns a rectangle and applies physics on it making it fall down until it is outside of the window.